+++ src/main/java/com/cleancoders/postgresqltranscations/repository/EmployeeRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -14,13 +15,13 @@ public interface EmployeeRepository extends JpaRepository<Employee, Long> {
 
     /** JPQL: returns employees whose salary exceeds the given threshold. */
     @Query("SELECT e FROM Employee e WHERE e.empSalary > :salary")
-    List<Employee> findBySalaryGreaterThan(@Param("salary") Long salary);
+    List<Employee> findBySalaryGreaterThan(@Param("salary") BigDecimal salary);
 
     /** Native SQL: same filter using the fully-qualified schema.table reference. */
     @Query(value = "SELECT * FROM company.employee WHERE emp_salary > :salary", nativeQuery = true)
-    List<Employee> findBySalaryGreaterThanNative(@Param("salary") Long salary);
+    List<Employee> findBySalaryGreaterThanNative(@Param("salary") BigDecimal salary);
 
     @Modifying
     @Query("DELETE FROM Employee e WHERE e.empSalary > ?1")
-    void deleteUsersBySalaryGreater(Long salary);
+    void deleteUsersBySalaryGreater(BigDecimal salary);
 }
